movies app
There are columns added for a video game collection, they can be empty if not necessary. There is a page added for viewing the Movies without pagination, accessible by the button "Full Table". The reviews can be added and seen in the table but are not stored in the database.